+++ app/src/main/java/com/aoztg/greengrim/presentation/ui/info/mycertification/MyCertificationFragment.kt
@@ -1,19 +1,13 @@
 package com.aoztg.greengrim.presentation.ui.info.mycertification
 
-import android.annotation.SuppressLint
-import android.graphics.Color
 import android.os.Bundle
 import android.view.View
-import android.widget.TextView
-import androidx.core.view.children
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
 import androidx.navigation.NavController
 import androidx.navigation.fragment.findNavController
 import com.aoztg.greengrim.MainNavDirections
 import com.aoztg.greengrim.R
-import com.aoztg.greengrim.databinding.CalendarDayLayoutBinding
-import com.aoztg.greengrim.databinding.CalendarHeaderBinding
 import com.aoztg.greengrim.databinding.FragmentMyCertificationBinding
 import com.aoztg.greengrim.presentation.base.BaseFragment
 import com.aoztg.greengrim.presentation.ui.BaseState
@@ -22,16 +16,8 @@ import com.aoztg.greengrim.presentation.ui.MonthState
 import com.aoztg.greengrim.presentation.ui.info.adapter.MyCertificationAdapter
 import com.aoztg.greengrim.presentation.ui.info.mycertification.MyCertificationViewModel.Companion.NEXT_PAGE
 import com.aoztg.greengrim.presentation.ui.main.MainViewModel
-import com.kizitonwose.calendar.core.CalendarDay
-import com.kizitonwose.calendar.core.CalendarMonth
-import com.kizitonwose.calendar.core.DayPosition
-import com.kizitonwose.calendar.core.daysOfWeek
-import com.kizitonwose.calendar.core.firstDayOfWeekFromLocale
-import com.kizitonwose.calendar.view.MonthDayBinder
-import com.kizitonwose.calendar.view.MonthHeaderFooterBinder
-import com.kizitonwose.calendar.view.ViewContainer
+import com.aoztg.greengrim.presentation.util.CustomCalendar
 import dagger.hilt.android.AndroidEntryPoint
-import java.time.DayOfWeek
 import java.time.LocalDate
 import java.time.YearMonth
 
@@ -43,19 +29,20 @@ class MyCertificationFragment :
     private val viewModel: MyCertificationViewModel by viewModels()
 
     private var selectedDate: LocalDate? = null
-    private var currentMonth: YearMonth = YearMonth.now()
-    private val today = LocalDate.now()
+    private lateinit var customCalendar: CustomCalendar
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
         parentViewModel.hideBNV()
         binding.vm = viewModel
-        binding.rvCertifications.adapter = MyCertificationAdapter()
-        viewModel.getCertificationDate()
+        initCustomCalendar()
         initStateObserver()
         initEventsObserver()
         setScrollEventListener()
+        setBtnClickListener()
+        binding.rvCertifications.adapter = MyCertificationAdapter()
+        viewModel.getCertificationDate()
     }
 
     private fun initStateObserver() {
@@ -102,45 +89,18 @@ class MyCertificationFragment :
                     is MyCertificationEvents.NavigateToCertificationDetail -> findNavController().toCertificationDetail(
                         it.certificationId
                     )
-
                     is MyCertificationEvents.ShowToastMessage -> showCustomToast(it.msg)
-                    is MyCertificationEvents.ShowCalendar -> initCalenderView()
+                    is MyCertificationEvents.ShowCalendar -> customCalendar.setDateWithDataList(
+                        viewModel.uiState.value.certificationDateList
+                    )
                     else -> {}
                 }
             }
         }
     }
 
-    @SuppressLint("SetTextI18n")
-    private fun initCalenderView() {
-        with(binding.calendarView) {
-
-            val startMonth = currentMonth.minusMonths(100)
-            val endMonth = currentMonth.plusMonths(100)
-            val firstDayOfWeek = firstDayOfWeekFromLocale()
-            val daysOfWeek = daysOfWeek()
-
-            setup(startMonth, endMonth, firstDayOfWeek)
-            scrollToMonth(currentMonth)
-
-            monthScrollListener = {
-                selectDate(it.yearMonth.atDay(1))
-                viewModel.scrollMonth(it.yearMonth)
-            }
-
-            binding.btnNextMonth.setOnClickListener {
-                currentMonth = currentMonth.plusMonths(1)
-                smoothScrollToMonth(currentMonth)
-            }
-
-            binding.btnPreviousMonth.setOnClickListener {
-                currentMonth = currentMonth.minusMonths(1)
-                smoothScrollToMonth(currentMonth)
-            }
-
-            configureBinders(daysOfWeek)
-        }
-
+    private fun yearMonthDatePickerConfirmListener(year:Int, month:Int){
+        customCalendar.yearMonthDatePickerConfirmListener(year, month)
     }
 
     private fun setScrollEventListener() {
@@ -168,92 +128,32 @@ class MyCertificationFragment :
 //                }
 //            }
 //        })
-
     }
 
-    private fun configureBinders(daysOfWeek: List<DayOfWeek>) {
-
-        class DayViewContainer(view: View) : ViewContainer(view) {
-            lateinit var day: CalendarDay
-            val textView = CalendarDayLayoutBinding.bind(view).calendarDayText
-
-            init {
-                view.setOnClickListener {
-                    if (day.position == DayPosition.MonthDate) {
-                        selectDate(day.date)
-                    }
-                }
-            }
+    private fun setBtnClickListener(){
+        binding.btnNextMonth.setOnClickListener{
+            customCalendar.goToNextMonth()
         }
 
-        binding.calendarView.dayBinder = object : MonthDayBinder<DayViewContainer> {
-
-            override fun create(view: View): DayViewContainer = DayViewContainer(view)
-
-            override fun bind(container: DayViewContainer, data: CalendarDay) {
-                container.day = data
-                val dateTv = container.textView
-
-                dateTv.text = data.date.dayOfMonth.toString()
-                if (data.position != DayPosition.MonthDate) {
-                    dateTv.visibility = View.GONE
-                } else {
-                    dateTv.visibility = View.VISIBLE
-                    when (data.date) {
-                        today -> {
-                            dateTv.setTextColor(Color.GRAY)
-                        }
-
-                        selectedDate -> {
-                            dateTv.setBackgroundResource(R.drawable.shape_calendar_selected)
-                            dateTv.setTextColor(Color.BLACK)
-                        }
-
-                        in viewModel.uiState.value.certificationDateList -> {
-                            dateTv.setBackgroundResource(R.drawable.shape_calendar_hasevent)
-                            dateTv.setTextColor(Color.WHITE)
-                        }
-
-                        else -> {
-                            dateTv.background = null
-                            dateTv.setTextColor(Color.WHITE)
-                        }
-                    }
-                }
-            }
-        }
-
-        class MonthViewContainer(view: View) : ViewContainer(view) {
-            val layout = CalendarHeaderBinding.bind(view).legendLayout.root
-        }
-
-        binding.calendarView.monthHeaderBinder =
-            object : MonthHeaderFooterBinder<MonthViewContainer> {
-                override fun create(view: View): MonthViewContainer = MonthViewContainer(view)
-                override fun bind(container: MonthViewContainer, data: CalendarMonth) {
-                    if (container.layout.tag == null) {
-                        container.layout.tag = data.yearMonth
-                        container.layout.children.map { it as TextView }
-                            .forEachIndexed { index, tv ->
-                                tv.text = daysOfWeek[index].name.first().toString()
-                            }
-                    }
-                }
-            }
-    }
-
-    private fun selectDate(date: LocalDate) {
-        if (selectedDate != date) {
-            val oldDate = selectedDate
-            oldDate?.let { binding.calendarView.notifyDateChanged(it) }
-            binding.calendarView.notifyDateChanged(date)
-            viewModel.selectDate(date)
+        binding.btnPreviousMonth.setOnClickListener{
+            customCalendar.goToPreviousMonth()
         }
     }
 
-    private fun yearMonthDatePickerConfirmListener(year: Int, month: Int) {
-        currentMonth = YearMonth.of(year, month)
-        binding.calendarView.scrollToMonth(currentMonth)
+    private fun initCustomCalendar(){
+        customCalendar = CustomCalendar(
+            binding.calendarView,
+            ::monthScrollListener,
+            ::dateSelectListener
+        )
+    }
+
+    private fun monthScrollListener(data: YearMonth){
+        viewModel.scrollMonth(data)
+    }
+
+    private fun dateSelectListener(data: LocalDate){
+        viewModel.selectDate(data)
     }
 
     private fun NavController.toCertificationDetail(certificationId: Int) {
